✨ Start both v1 and v2 Docker containers
Start separate Docker containers for both Influx v1 and v2.

The v2 Docker containers aren't yet being published to DockerHub but are available on quay.io, so we're pulling from there for now.

While v2 is in beta, it uses port 9999 by default, so that's what we're using here.  Once v2 is released, that will change and we'll have to adapt to that in a future PR.

The v2 Docker container does not currently auto-provision itself using environment variables, so we're running an explicit curl command to do that.

All of the curl commands write to separate error log files so that no potentially-helpful debugging information gets lost.

Also, added the error-*.log files to `.gitignore` to keep them from showing up as changed.

Co-authored-by: Iris Scholten <iris@influxdata.com>